chor(GiniHealthSDK): Fix bundle in storyboard.
PIA-1981
