+++ HealthSDK/GiniHealthSDK/Documentation/source/Installation.md
@@ -24,7 +24,7 @@ dependencies: [
 ## Manually
 
 If you prefer not to use a dependency management tool, you can integrate the Gini Health SDK into your project manually.
-To do so drop the GiniPayBusiness (classes and assets) folder into your project and add the files to your target.
+To do so drop the GiniHealth (classes and assets) folder into your project and add the files to your target.
 
 Xcode will automatically check your project for swift files and will create an autogenerated import header for you.
 Use this header in an Objective-C project by adding
