+++ HealthSDK/GiniHealthSDK/Sources/GiniHealthSDK/Core/PaymentReviewViewController.swift
@@ -681,23 +681,3 @@ extension PaymentReviewViewController {
         present(alertController, animated: true, completion: nil)
     }
 }
-
-class GiniCustomButton: UIButton {
-    var disabledBackgroundColor: UIColor? = .gray
-    var defaultBackgroundColor: UIColor? {
-        didSet {
-            backgroundColor = defaultBackgroundColor
-        }
-    }
-    override public var isEnabled: Bool {
-        didSet {
-            self.backgroundColor = isEnabled ? defaultBackgroundColor : disabledBackgroundColor
-        }
-    }
-    
-    override var isHighlighted: Bool {
-        didSet {
-            self.alpha = isHighlighted ? 0.5 : 1
-        }
-    }
-}
